Restructure and name changing stuff to make everything more readable
