+++ Embedder/src/utils.ts
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import { encode } from 'gpt-tokenizer/cjs/model/gpt-4-32k-0314';
 
 export async function getBase64(url: string): Promise<string> {
     return axios
@@ -17,4 +18,13 @@ export function secondsToTimestamp(convertSeconds: number): string {
     return hours.toString().padStart(2, '0')
         + ':' + minutes.toString().padStart(2, '0')
         + ':' + seconds.toString().padStart(2, '0');
+}
+
+export function trimTextByTokenSize(text: string, maxTokens: number): string {
+    let editText = text;
+    while (encode(editText).length >= maxTokens) {
+        editText = editText.slice(0, -1);
+    }
+
+    return editText;
 }
